General cleanup

+++ test/t/004unknown.t
@@ -11,10 +11,10 @@ my $text;
 # setup
 $ENV{DEBUG} = 1;   # this enables the check_MV.pl script to use data files in the local directory
 #`/usr/bin/cp -f config.txt.critical config.txt` ;
-#`cp -f exproxy.hst.warning ezproxy.hst`;
+#`cp -f ezproxy.hst.warning ezproxy.hst`;
 
 # The Test
-`/bin/rm config.txt`;
+`/bin/rm config.txt` if (-e "config.txt");
 `../check_MV.pl`; 
 $return_value = $? >> 8;
 ok ($return_value == 3, "script should return a three, when it's unable to monitor the number of VirtualHosts: $return_value");
